refactor(core): split graph asset helpers
